+++ BuildMasterAutomation/Functions/Invoke-BMRestMethod.ps1
@@ -133,7 +133,7 @@ function Invoke-BMRestMethod
                     'X-ApiKey' = $Session.ApiKey;
                 }
 
-    #$DebugPreference = 'Continue'
+    # $DebugPreference = 'Continue'
     Write-Debug -Message ('{0} {1}' -f $Method.ToString().ToUpperInvariant(),($uri -replace '\b(API_Key=)([^&]+)','$1********'))
     if( $ContentType )
     {